Rust Compiler likes me. Needs unit tests for Board

+++ src/piece_move.rs
@@ -105,30 +105,30 @@ impl BitMove {
     pub fn is_capture(&self) -> bool    { ((&self.data & CP_MASK) >> 14) == 1 }
 
     pub fn is_promo(&self) -> bool { ((&self.data & PR_MASK) >> 15) == 1 }
-    pub fn get_dest(&self) -> u16       { (&self.data & DST_MASK) >> 6 }
-    pub fn get_src(&self) -> u16        { (&self.data & SRC_MASK)}
+    pub fn get_dest(&self) -> u8       { ((&self.data & DST_MASK) >> 6) as u8 }
+    pub fn get_src(&self) -> u8        { (&self.data & SRC_MASK) as u8}
     pub fn is_castle(&self) -> bool     { ((&self.data & FLAG_MASK) >> 13) == 1 }
     pub fn is_king_castle(&self) -> bool {((&self.data & FLAG_MASK) >> 12) == 2}
     pub fn is_queen_castle(&self) -> bool {((&self.data & FLAG_MASK) >> 12) == 3}
     pub fn is_en_passant(&self) -> bool { (&self.data & FLAG_MASK) >> 12 == 5}
     pub fn is_double_push(&self) -> (bool, u8) {
-        let is_double_push: u8 = (&self.data & FLAG_MASK);
+        let is_double_push: u8 = (&self.data & FLAG_MASK) as u8;
         match is_double_push {
-            1 => (true, self.get_dest()),
+            1 => (true, self.get_dest() as u8),
             _ => (false, 64)
         }
     }
     pub fn dest_row(&self) -> u8 {
-        ((self.data & DST_MASK) >> 6) / 8
+        ((self.data & DST_MASK) >> 6) as u8 / 8
     }
     pub fn dest_col(&self) -> u8 {
-        ((self.data & DST_MASK) >> 6) % 8
+        ((self.data & DST_MASK) >> 6) as u8 % 8
     }
     pub fn src_row(&self) -> u8 {
-        (self.data & SRC_MASK) / 8
+        (self.data & SRC_MASK) as u8 / 8
     }
     pub fn src_col(&self) -> u8 {
-        (self.data & SRC_MASK) % 8
+        (self.data & SRC_MASK) as u8 % 8
     }
     pub fn promo_piece(&self) -> Piece {
         match (self.data >> 12) & 0b0011 {
